Modified the generic serve action and http corpse
The way nataniev was calling vessels before, it was instantiating a new instance
of each vessel every time it called an action on it. Since this is no longer
happening, much of the state is being preserved. One of the things that was
happening was that scripts, links, and footers were getting endlessly appended
to the http corpse since every time it is served the corpse is built. Blanking
out the metas, links, scripts and footers attributes seemed easier than refactoring
every vessel's invoke.rb to not build every time.