MemoryArray refresh.
Deprecated methods that write the memory file and then perform re-render 
in favor of new ones that only perform the file writes. Added new 
functionality to MemoryArray to handle memory updates in ways that do 
not require re-rendering the entire memory file. These changes should 
also allow to move file writes to threads so that they don't block 
execution of the rest of the code.